kada-mandiya-analytics/requirements.txt
apscheduler
black
fastapi
loguru
matplotlib
networkx
numpy
pandas
plotly
pydantic>=2,<3
pyodbc
pytest
python-dotenv
ruff
scipy
seaborn
sqlalchemy>=2,<3
streamlit
tenacity
uvicorn
